JCU Organisational Change support
These Mint changes support the altered JCU organisational structure. For
users of the Dashboard, only current Groups are presented. (where
activeto = “”) If an Administrator was to edit a published record,
current Groups (Affiliations)  are but the description is tagged
with the group ID. This approach was taken so a distinction between
current and old groups can be distinguished by Administrator users. The
new fields Old_ID, Active_From and Active_To were added.